feat: Q-Learning adaptive engine, ESC1/ESC6, zero clippy warnings, full audit
Major changes:
- Q-Learning reinforcement learning attack engine (overthrone-pilot, qlearn feature)
- ADCS ESC1 exploiter (204 lines, SAN UPN abuse)
- ADCS ESC6 exploiter (200 lines, EDITF_ATTRIBUTESUBJECTALTNAME2)
- DSRM pass-the-hash backdoor wired
- WMI DCOM/EPM endpoint mapper
- CLI PDF output wired to scribe::pdf::render()
- TUI crawler integration wired
- C2 implant deploy CLI wired
- WinRM Windows real output collection
- WASM plugin state persistence + manifest parsing fixed
- Zero clippy warnings across entire workspace (-D warnings)
- 222 core + 8 pilot = 230 tests passing, 0 failures
- README/IMPLEMENTATION_STATUS/TODO updated with accurate status

Cleanup:
- Removed output.txt, stale duplicate overthrone-cli/, internal PLAN docs
- Removed unused test_pfx.rs binary, setup.ps1 scaffolding script
- Removed redundant uncompressed wordlist (only .zst used)
- Updated .gitignore for all above

+++ crates/overthrone-forge/src/shadow_credentials.rs
@@ -390,10 +390,10 @@ fn generate_key_pair(key_size: u16) -> Result<KeyPair> {
 
     // Set validity period
     let now = chrono::Utc::now();
-    params.not_before = rcgen::date_time_ymd(now.year() as i32, now.month() as u8, now.day() as u8);
+    params.not_before = rcgen::date_time_ymd(now.year(), now.month() as u8, now.day() as u8);
     let expiry = now + chrono::Duration::days(365);
     params.not_after = rcgen::date_time_ymd(
-        expiry.year() as i32,
+        expiry.year(),
         expiry.month() as u8,
         expiry.day() as u8,
     );
